+++ SourceCode/Database/dbo/Tables/Document.sql
@@ -1,6 +1,6 @@
 ﻿CREATE TABLE [dbo].[Document] (
     [Id]               INT                IDENTITY (1, 1) NOT NULL,
-    [FileExtension]    NCHAR (5)          NOT NULL,
+    [FileExtension]    NVARCHAR(6)        NOT NULL,
     [ContentType]      NVARCHAR (50)      NULL,
     [Content]          VARBINARY (MAX)    NULL,
     [LastModifiedTime] DATETIMEOFFSET (7) NULL,