+++ SourceCode/Services/Implementations/StationCustomerService.cs
@@ -21,6 +21,26 @@ public class StationCustomerService
         return Array.Empty<StationCustomer>();
     }
 
+    public async Task<IEnumerable<StationCustomerCargo>> GetGetCustomerCargoAsync(ClaimsPrincipal? principal, int cargoId, int countryId = 0)
+    {
+        if (principal.IsAuthenticated())
+        {
+            using var dbContext = Factory.CreateDbContext();
+            return await dbContext.StationCustomerCargos
+                .Where(scc => scc.CargoId == cargoId && (countryId == 0 || scc.StationCustomer.Station.Region.CountryId == countryId))
+                .Include(c => c.Cargo)
+                .Include(scc => scc.Direction)
+                .Include(scc => scc.QuantityUnit)
+                .Include(sc => sc.StationCustomer)
+                    .ThenInclude(s => s.Station)
+                    .ThenInclude(s => s.Region)
+                    .ThenInclude(r => r.Country)
+                .ToReadOnlyListAsync();
+        }
+        return Array.Empty<StationCustomerCargo>();
+    }
+
+
     public async Task<StationCustomer?> FindByIdAsync(ClaimsPrincipal? principal, int id)
     {
         if (principal.IsAuthenticated())
@@ -32,6 +52,7 @@ public class StationCustomerService
         }
         return null;
     }
+
     public Task<(int Count, string Message, StationCustomer? Entity)> SaveAsync(ClaimsPrincipal? principal, int stationId, StationCustomer entity) =>
         SaveAsync(principal, stationId, entity, principal.AsModuleOwnershipRef());